use composition for App components
