+++ .env
@@ -1,0 +1,1 @@
+API_URL=http://localhost:8000/src/api
