+++ src/components/Footer.jsx
@@ -1,0 +1,40 @@
+export default function Footer() {
+    return <footer>
+    <span className="hidden lg:inline-flex fixed right-0 bottom-0 p-2 mt-auto text-sm items-center text-gray-500 font-thins">
+      Sya v1.1-ALPHA
+      <a
+        title="Powered by sycatle.dev"
+        href="https://sycatle.dev"
+        target="__blank"
+        className=" ml-1"
+      >
+        <svg
+          className="h-10 w-10 fill-black dark:fill-white hover:scale-105 duration-200"
+          xmlns="http://www.w3.org/2000/svg"
+          xmlnsXlink="http://www.w3.org/1999/xlink"
+          zoomAndPan="magnify"
+          viewBox="0 0 375 374.999991"
+          preserveAspectRatio="xMidYMid meet"
+          version="1.0"
+          data-arp-injected="true"
+        >
+          <defs>
+            <clipPath>
+              <path
+                d="M 40.511719 40.511719 L 334.511719 40.511719 L 334.511719 334.511719 L 40.511719 334.511719 Z M 40.511719 40.511719 "
+                clipRule="nonzero"
+              ></path>
+            </clipPath>
+          </defs>
+          <g clipPath="url(#f7eadf1f08)">
+            <path
+              d="M 261.558594 142.507812 C 257.972656 136.644531 253.6875 131.136719 248.789062 126.324219 C 233.128906 110.667969 211.4375 100.957031 187.554688 100.957031 C 163.675781 100.957031 141.980469 110.667969 126.324219 126.324219 C 116.964844 135.683594 109.792969 147.144531 105.417969 159.914062 L 201.902344 178.808594 L 208.8125 180.121094 L 285.964844 195.167969 L 296.546875 197.265625 C 294.1875 223.59375 282.554688 247.214844 264.972656 264.882812 C 245.203125 284.652344 217.820312 296.898438 187.644531 296.898438 C 148.105469 296.898438 111.980469 281.851562 84.773438 257.273438 L 84.773438 257.183594 C 107.867188 291.300781 146.355469 311.769531 187.644531 311.769531 C 221.933594 311.769531 252.988281 297.859375 275.46875 275.378906 C 297.949219 252.898438 311.855469 221.847656 311.855469 187.554688 C 311.855469 185.019531 311.769531 182.570312 311.683594 180.121094 L 296.546875 177.148438 L 273.019531 172.511719 L 195.253906 157.289062 L 331.539062 157.289062 C 333.550781 167 334.601562 177.148438 334.601562 187.46875 C 334.601562 228.058594 318.15625 264.796875 291.5625 291.386719 C 264.972656 317.980469 228.230469 334.425781 187.644531 334.425781 C 147.054688 334.425781 110.316406 317.980469 83.726562 291.386719 C 67.453125 275.117188 54.945312 255 47.683594 232.519531 C 67.628906 232.519531 87.574219 232.519531 107.519531 232.519531 L 113.640625 232.519531 C 117.226562 238.378906 121.511719 243.890625 126.410156 248.699219 C 142.070312 264.359375 163.761719 274.066406 187.644531 274.066406 C 211.523438 274.066406 233.21875 264.359375 248.875 248.699219 C 258.234375 239.339844 265.40625 227.882812 269.695312 215.109375 L 179.070312 197.441406 L 166.300781 194.902344 L 89.148438 179.859375 L 78.5625 177.757812 C 80.839844 151.429688 92.558594 127.8125 110.140625 110.140625 C 129.910156 90.371094 157.289062 78.125 187.46875 78.125 C 227.007812 78.125 263.132812 93.171875 290.339844 117.753906 C 285.878906 111.191406 280.890625 105.15625 275.378906 99.644531 C 252.898438 77.164062 221.847656 63.253906 187.554688 63.253906 C 153.265625 63.253906 122.214844 77.164062 99.730469 99.644531 C 77.25 122.125 63.34375 153.179688 63.34375 187.46875 C 63.34375 190.003906 63.429688 192.453125 63.519531 194.902344 L 78.652344 197.878906 L 102.183594 202.515625 L 179.945312 217.734375 C 134.546875 217.734375 89.0625 217.734375 43.660156 217.734375 C 41.648438 208.023438 40.601562 197.878906 40.601562 187.554688 C 40.601562 146.96875 57.046875 110.230469 83.636719 83.636719 C 110.230469 57.046875 146.96875 40.601562 187.554688 40.601562 C 228.144531 40.601562 264.882812 57.046875 291.476562 83.636719 C 307.746094 99.90625 320.253906 120.027344 327.515625 142.507812 C 305.558594 142.507812 283.515625 142.507812 261.558594 142.507812 Z M 261.558594 142.507812 "
+              fillOpacity="1"
+              fillRule="nonzero"
+            ></path>
+          </g>
+        </svg>
+      </a>
+    </span>
+  </footer>
+}
