+++ src/components/Security/Identifier.jsx
@@ -1,0 +1,16 @@
+export default function Identifier({ setter }) {
+  return (
+    <label htmlFor="identifier" className="flex flex-col space-y-2">
+      <span className="text-lg font-semibold">
+        Pseudonyme / Adresse électronique
+      </span>
+      <input
+        required
+        onChange={(e) => setter(prevState => ({...prevState, username: e.target.value}))}
+        type="text"
+        name="identifier"
+        className="rounded-lg shadow focus:border-b-2 border-[#6fb463] bg-white dark:bg-black p-2 outline-none"
+      />
+    </label>
+  );
+}
